Initial Commit
The radiation course I'm auditing has a fun incremental project alongside the assignments. This is a relativistic particle solver, including external electric and magnetic forces.

The initial commit of this code will not be its final form. For now, it will likely be confusing to run. Many edits are to come, and I will make it far more user-friendly by the time I've finished.

I did not use a Boris pusher as my solver. I made my algorithm using a scheme similar to a suggestion by my undergraduate supervisor for an unrelated project, which I corrected and made more robust. We called it a 'double leapfrog'. I'm not sure if this is popular or widespread, but either way, it's fun!